+++ packages/brick-kit/src/developHelper.ts
@@ -1,7 +1,12 @@
 import { asyncProcessBrick } from "@easyops/brick-utils";
 import { BrickConf } from "@easyops/brick-types";
 import { brickTemplateRegistry } from "./core/TemplateRegistries";
-import { LocationContext, mountTree, unmountTree } from "./core/exports";
+import {
+  LocationContext,
+  mountTree,
+  unmountTree,
+  _dev_only_getBrickPackages
+} from "./core/exports";
 
 export const developHelper = {
   asyncProcessBrick(brickConf: BrickConf): Promise<void> {
@@ -9,5 +14,6 @@ export const developHelper = {
   },
   LocationContext,
   mountTree,
-  unmountTree
+  unmountTree,
+  getBrickPackages: _dev_only_getBrickPackages
 };