+++ packages/browserslist-config-next/development.js
@@ -1,0 +1,2 @@
+// Ref https://github.com/browserslist/browserslist#shareable-configs
+module.exports = ["last 2 Chrome versions"];