+++ packages/browserslist-config-next/index.js
@@ -1,0 +1,2 @@
+// Ref https://github.com/browserslist/browserslist#shareable-configs
+module.exports = require("./production");