+++ packages/create-next-repo/template/.size-limit.js
@@ -1,0 +1,10 @@
+const { sizeLimit } = require("@easyops/build-config-factory");
+
+module.exports = sizeLimit({
+  bricks: {
+    "*": "50 KB"
+  },
+  templates: {
+    "*": "5 KB"
+  }
+});