+++ packages/create-next-repo/template/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  presets: ["@easyops/babel-preset-next"]
+};