+++ packages/webpack-config-factory/src/bricks/webpack.common.factory.js
@@ -41,7 +41,8 @@ const getImageLoaderOptions = distPublicPath => ({
       options: {
         name: "assets/[name].[hash:8].[ext]",
         limit: 8192,
-        publicPath: distPublicPath
+        publicPath: distPublicPath,
+        esModule: false
       }
     }
   ]
@@ -58,6 +59,7 @@ module.exports = ({ scope = "bricks" } = {}) => {
   const imageLoaderOptions = getImageLoaderOptions(distPublicPath);
 
   const packageJson = require(path.join(cwdDirname, "package.json"));
+  const packageName = packageJson.name.split("/")[1];
   const dll = Object.keys(packageJson.devDependencies).filter(name =>
     name.startsWith("@dll/")
   );
@@ -176,8 +178,9 @@ module.exports = ({ scope = "bricks" } = {}) => {
     },
     plugins: [
       scope === "templates"
-        ? new ScanTemplatesPlugin()
+        ? new ScanTemplatesPlugin(packageName)
         : new ScanCustomElementsPlugin(
+            packageName,
             dll.map(name => name.substr("@dll/".length))
           ),
       new CleanWebpackPlugin(),