+++ dll/d3/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.1](https://git.easyops.local/anyclouds/next-core/compare/@dll/d3@1.0.0...@dll/d3@1.0.1) (2019-11-27)
+
+**Note:** Version bump only for package @dll/d3
+
 # [1.0.0](https://git.easyops.local/anyclouds/next-core/compare/@dll/d3@0.3.5...@dll/d3@1.0.0) (2019-11-03)
 
 **Note:** Version bump only for package @dll/d3