+++ packages/brick-utils/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-utils@1.3.0...@easyops/brick-utils@1.3.1) (2019-11-27)
+
+**Note:** Version bump only for package @easyops/brick-utils
+
+# [1.3.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-utils@1.2.2...@easyops/brick-utils@1.3.0) (2019-11-26)
+
+### Features
+
+- **brick-utils:** 构件事件支持设置目标的属性 ([a0d9b9b](https://git.easyops.local/anyclouds/next-core/commits/a0d9b9b))
+
 ## [1.2.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-utils@1.2.1...@easyops/brick-utils@1.2.2) (2019-11-13)
 
 ### Bug Fixes