+++ packages/custom-antd-styles/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.6.0...@easyops/custom-antd-styles@0.7.0) (2019-11-27)
+
+### Features
+
+- checkbox style modification ([65dafb4](https://git.easyops.local/anyclouds/next-core/commits/65dafb4))
+- tabs style modification ([8cf83b3](https://git.easyops.local/anyclouds/next-core/commits/8cf83b3))
+
+# [0.6.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.5.0...@easyops/custom-antd-styles@0.6.0) (2019-11-26)
+
+### Features
+
+- modify small size button height ([4136685](https://git.easyops.local/anyclouds/next-core/commits/4136685))
+
+# [0.5.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.4.0...@easyops/custom-antd-styles@0.5.0) (2019-11-21)
+
+### Features
+
+- gradient button style ([8adc20d](https://git.easyops.local/anyclouds/next-core/commits/8adc20d))
+
+# [0.4.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.3.0...@easyops/custom-antd-styles@0.4.0) (2019-11-21)
+
+### Features
+
+- set card-head-padding to 12px ([a750ae0](https://git.easyops.local/anyclouds/next-core/commits/a750ae0))
+
+# [0.3.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.2.0...@easyops/custom-antd-styles@0.3.0) (2019-11-20)
+
+### Features
+
+- add link color style ([ebdf813](https://git.easyops.local/anyclouds/next-core/commits/ebdf813))
+- add tag component style ([c09075a](https://git.easyops.local/anyclouds/next-core/commits/c09075a))
+
+# [0.2.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.1.1...@easyops/custom-antd-styles@0.2.0) (2019-11-19)
+
+### Features
+
+- **custom-style:** custom style of antd alert, refs BRICK_STORE-642 ([4846378](https://git.easyops.local/anyclouds/next-core/commits/4846378))
+
 ## [0.1.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/custom-antd-styles@0.1.0...@easyops/custom-antd-styles@0.1.1) (2019-11-12)
 
 ### Bug Fixes